pyproject.toml
[project]
name = "assitment-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "huggingface-hub==0.16.4",
    "python-telegram-bot==20.7",
    "sentence-transformers==2.2.2",
    "torch>=2.0.0",
    "transformers==4.30.2",
]
